finish tweak crearPedido backend - psql se confunde si se insertan las PKs a mano :,v piensa que la secuencia aun esta en 1

+++ Laboratorio1/backend/src/main/java/com/example/tbd_lab1/entities/ProductoEntity.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class ProductoEntity {
     private Long idProducto;
     private String nombreProducto;
-    private float precio;
+    private Integer precio;
     private String categoria;
     private boolean requiereReceta;
     private String imageUrl;
